+++ lib/puppy.js
@@ -2,7 +2,7 @@ const fm = require('front-matter');
 const { concat, through } = require('mississippi');
 const stream = require('stream');
 const vfs = require('vinyl-fs');
-const { loadPages, loadData } = require('./utils');
+const { loadPages, loadData } = require('./context');
 
 const defaultOptions = {
   pages: './pages/**/*',
